tests/requirements.txt
trame
trame-client[test]
trame-vuetify
vtk==9.4.2
pyvista[all]==0.45.2
pytest-asyncio
pytest-playwright
coverage
